+++ api/src/pdc/services/policy/engine/policies/index.ts
@@ -9,7 +9,7 @@ import { LaRochelle20232024 } from './20230101_LaRochelle';
 import { Lannion } from './20230501_Lannion';
 import { LannionTregor2024 } from './20240401_LannionTregor';
 import { Laval } from './20220412_LavalAgglo';
-import { MetropoleSavoie } from './MetropoleSavoie';
+import { MetropoleSavoie } from './20230124_MetropoleSavoie';
 import { Montpellier } from './Montpellier';
 import { Mrn } from './Mrn';
 import { NantesMetropole2024 } from './NantesMetropole2024';
